+++ R/manystates-states.R
@@ -1,15 +1,15 @@
-#' States database documentation
+#' states database documentation
 #'
 #' @format The states database is a list that contains the
 #' following 3 datasets: GW, ISD, COW.
 #' For more information and references to each of the datasets used,
-#' please use the `data_source()`, `data_contrast()` functions.
+#' please use the `data_source()` and `data_contrast()` functions.
 #'\describe{
 #' \item{GW: }{A dataset with 216 observations and the following
-#' 5 variables: ID, Beg, End, COW_Nr, Label.}
+#' 5 variables: GW_ID, Beg, End, COW_Nr, Label.}
 #' \item{ISD: }{A dataset with 362 observations and the following
-#' 7 variables: ID, Beg, End, COW_Nr, Label, Micro, New.State.}
+#' 7 variables: ISD_ID, Beg, End, COW_Nr, Label, Micro, New.State.}
 #' \item{COW: }{A dataset with 243 observations and the following
-#' 5 variables: ID, Beg, End, COW_Nr, Label.}
+#' 5 variables: COW_ID, Beg, End, COW_Nr, Label.}
 #' }
  "states"