+++ data-raw/states/COW/prepare-COW.R
@@ -11,14 +11,14 @@ COW <- readr::read_csv("data-raw/states/COW/states2016.csv")
 # formats of the 'COW' object until the object created
 # below (in stage three) passes all the tests.
 COW <- tibble::as_tibble(COW) %>%
-  qData::transmutate(ID = stateabb,
-                     Beg = qCreate::standardise_dates(lubridate::as_date(paste(styear, stmonth, stday, sep = "-"))),
-                     End = qCreate::standardise_dates(lubridate::as_date(paste(endyear, endmonth, endday, sep = "-"))),
-                     Label = qCreate::standardise_titles(statenme),
-                     COW_Nr = qCreate::standardise_titles(as.character(ccode))) %>%
-  dplyr::select(COW_Nr, ID, Beg, End, Label) %>%
-  dplyr::relocate(ID, Beg, End, COW_Nr, Label) %>%
-  dplyr::arrange(Beg, ID)
+  manydata::transmutate(COW_ID = stateabb,
+                     Beg = manypkgs::standardise_dates(lubridate::as_date(paste(styear, stmonth, stday, sep = "-"))),
+                     End = manypkgs::standardise_dates(lubridate::as_date(paste(endyear, endmonth, endday, sep = "-"))),
+                     Label = manypkgs::standardise_titles(statenme),
+                     COW_Nr = manypkgs::standardise_titles(as.character(ccode))) %>%
+  dplyr::select(COW_ID, COW_Nr, Beg, End, Label) %>%
+  dplyr::relocate(COW_ID, Beg, End, COW_Nr, Label) %>%
+  dplyr::arrange(Beg, COW_ID)
 # We know that COW data for "old" states is set to 1816-01-01 by default.
 # This is a rather uncretain date, that is, the dataset considers them states
 # on 1st January 1816, but they may have been established (much) earlier.
@@ -26,14 +26,14 @@ COW <- tibble::as_tibble(COW) %>%
 # which is designed to deal with date uncertainty
 COW <- COW %>% dplyr::mutate(Beg = messydates::as_messydate(ifelse(Beg <= "1816-01-01", messydates::on_or_before(Beg), Beg)),
                              End = messydates::as_messydate(ifelse(End >= "2016-12-31", messydates::on_or_after(End), End)))
-# qData and qCreate include several other
+# manydata and manypkgs include several other
 # functions that should help cleaning and
 # standardizing your data.
 # Please see the vignettes or website for more details.
 
 # Stage three: Connecting data
 # Next run the following line to make COW available within the qPackage.
-qCreate::export_data(COW, database = "states",
+manypkgs::export_data(COW, database = "states",
                      URL = "https://correlatesofwar.org/data-sets/state-system-membership")
 
 # This function also does two additional things.