+++ data-raw/states/ISD/isd_actor.R
@@ -1,5 +1,5 @@
 library(tidyverse)
-library(qData)
+library(manydata)
 library(countrycode)
 # Get Griffiths/Butcher ISD data ####
 isd_actor <- read.csv("data-raw/states/ISD_Version1_Dissemination.csv",